fix: move static files to public and express to api folder
