+++ public/js/randomMsg.js
